backend/pyproject.toml
[project]
name = "app"
version = "0.0.4"
description = ""
authors = [{ name = "Anderson Resende", email = "anderson@vinta.com.br" }]
requires-python = ">=3.12,<3.13"
readme = "README.md"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "asyncpg>=0.29.0",
    "fastapi-users[sqlalchemy]>=13.0.0",
    "pydantic-settings>=2.5.2",
    "fastapi-mail>=1.4.1",
    "openai>=1.99.6",
    "pillow>=10.4.0",
    "uvicorn[standard]>=0.35.0",
    "authlib>=1.6.1",
    "pyjwt>=2.10.1",
    "sqlalchemy[asyncio]>=2.0.42",
    "httpx>=0.28.1",
    "python-multipart>=0.0.20",
    "alembic>=1.16.4",
    "psycopg[binary]>=3.2.9",
    "passlib[bcrypt]>=1.7.4",
    "jinja2>=3.1.6",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "torch>=2.8.0",
    "transformers>=4.55.2",
    "soundfile>=0.13.1",
    "torchaudio>=2.8.0",
    "onnxruntime>=1.22.1",
]

[dependency-groups]
dev = [
    "pre-commit>=3.4.0",
    "ruff>=0.1.0",
    "watchdog>=5.0.3",
    "python-dotenv>=1.0.1",
    "pytest>=8.3.3",
    "pytest-mock>=3.14.0",
    "mypy>=1.13.0",
    "coveralls>=4.0.1",
    "alembic>=1.14.0",
    "pytest-asyncio>=0.24.0",
    "mkdocs-material[imaging]>=9.6.9",
]

[tool.uv]
package = false

[tool.hatch.build.targets.sdist]
include = ["commands"]

[tool.hatch.build.targets.wheel]
include = ["commands"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
